Updating files to reflect the version in the AKNNISS repo

+++ app/LSH/hashFunctions/HashFunction.scala
@@ -5,5 +5,6 @@ package LSH.hashFunctions
   */
 @SerialVersionUID(100L)
 abstract class HashFunction(k:Int) extends Serializable {
-  def apply(v:Vector[Float]) : String
+  def apply(v:Vector[Double]) : String
 }
+
