+++ app/LSH/hashFunctions/Hyperplane.scala
@@ -1,6 +1,6 @@
 package LSH.hashFunctions
 
-import LSH.tools.Distance
+import tools.Distance
 
 import scala.collection.mutable.ArrayBuffer
 import scala.util.Random
@@ -10,7 +10,7 @@ import scala.util.Random
   */
 class Hyperplane(k:Int) extends HashFunction(k) {
 
-  var hyperPlanes = new ArrayBuffer[Vector[Float]]
+  var hyperPlanes = new ArrayBuffer[Vector[Double]]
   val rnd = new Random(System.currentTimeMillis())
 
   // Initialize k random hyperplanes for H[i]
@@ -18,7 +18,7 @@ class Hyperplane(k:Int) extends HashFunction(k) {
     hyperPlanes += generateRandomV(4096)
   }
 
-  def apply(v: Vector[Float]): String = {
+  def apply(v: Vector[Double]): String = {
     var sb : StringBuilder = new StringBuilder
     for(h <- hyperPlanes) {
       sb ++= hash(v,h).toString
@@ -26,16 +26,17 @@ class Hyperplane(k:Int) extends HashFunction(k) {
     // TODO Dont use string
     sb.toString()
   }
-  def hash(v: Vector[Float], randomV: Vector[Float]): Int = {
+  def hash(v: Vector[Double], randomV: Vector[Double]): Int = {
     if (Distance.dotProduct(v, randomV) > 0) 1 else 0
   }
 
   // TODO check the seq structure
-  def generateRandomV(size: Int): Vector[Float] = {
-    val buf = new ArrayBuffer[Float]
+  def generateRandomV(size: Int): Vector[Double] = {
+    val buf = new ArrayBuffer[Double]
     for (i <- 0 until size)
       buf += (if (rnd.nextGaussian() < 0) -1 else 1)
 
     buf.toVector
   }
 }
+
