+++ app/LSH/structures/HashTable.scala
@@ -1,7 +1,6 @@
 package LSH.structures
 
 import LSH.hashFunctions.HashFunction
-
 import scala.collection.mutable
 
 /**
@@ -10,7 +9,7 @@ import scala.collection.mutable
 @SerialVersionUID(100L)
 class HashTable(f:() => HashFunction) extends Serializable {
   // internal Mutable HashMap
-  val table = new mutable.HashMap[String, List[(String, Vector[Float])]]()
+  val table = new mutable.HashMap[String, List[(String, Vector[Double])]]()
   // internal Hash function
   val hf = f()
 
@@ -18,7 +17,7 @@ class HashTable(f:() => HashFunction) extends Serializable {
     * Insert vector
     * @param v vector to be inserted into internal hashmap
     */
-  def +=(v:(String, Vector[Float])) : Unit = {
+  def +=(v:(String, Vector[Double])) : Unit = {
     val key = hf(v._2)
     val value = {
       if(table.contains(key)) table(key)++List(v)
@@ -31,9 +30,10 @@ class HashTable(f:() => HashFunction) extends Serializable {
     * @param v a query point
     * @return a list of vectors with same key as v
     */
-  def query(v:Vector[Float]) : Stream[(String, Vector[Float])] = {
+  def query(v:Vector[Double]) : Stream[(String, Vector[Double])] = {
     val key = hf(v)
     table(key).toStream
   }
 
 }
+
